requirements/base.txt
Django==2.0.6
django-environ==0.4.4
djangorestframework==3.8.2
django-filter==1.1.0
